A few theme fixes and some extra authenticity added
Now default buttons in dialogs have a black border around them
as an indicator that they are default buttons.

The `start-here` icon is now symbolic, so that it is colored
depending upon the current theme and the variant of the theme.

A new style rule file has been added for LightDM.