Theme the Caja file manager in Icon / Compact view correctly
Despite the name of the file manager's mode, Icon View (in Caja) does not
use a GtkIconView like one would expect.  It uses a purpose-built,
file-manager-specific widget called CajaIconContainer.

When one renames a file while Caja is in Icon View or Compact View mode
(both use CajaIconContainer), another specialty, non-standard widget is
used: the EelEditableLabel.  The editable label widget is supposed to
look like a GtkEntry widget, but they are not quite implemented the same,
and as such theme designers need to add style declarations for the editable
label separate from the declarations for GtkEntry.

Gtk-Theme-Raleigh did not theme EelEditableLabels in the past, but this
commit now themes the labels like GtkEntry widgets.