Update README, "zebra stripes" notice, and screenshots
The screenshots were 17 months old.  Time for an update maybe?
I also updated the README to note that the new screenshots were taken
on a MATE version 1.26.2 desktop, as opposed to the old screenshots
which were taken on a 1.26.0 desktop.

Now the zebra stripes notice contains a "TL;DR" notice for those who
just want zebra stripes, not a few pages of boring text about why
it's so hard to get the zebra stripes right.

Maybe next commit I'll get around to providing XFCE screenshots; up
to this point, I've wanted to, but there used to be a few icons here
and there which were missing.  That's not good publicity for a theme!